requirements.txt
fastapi[standard]===0.115.6
itsdangerous==2.2.0
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
pymongo==4.10.1
pyjwt==2.10.1
python-jose==3.3.0
requests==2.32.3